Rename myChatAppm.png to Ichocho Fitness App .png
Ichocho Fitness App is the mobile application that has been proposed and it is targeting individuals interested in fitness and wellness. Ichocho Fitness App purpose is to provide a complete platform for users to set and attain fitness goals, access personalized workout routines, track progress, and connect with like-minded fitness enthusiasts. The app centers on creating a supportive community, nurtures motivation and offers a user friendly interface for fitness experience.